Notifications modal (#37)
* Implement notifications modal

Basic structure

* Change number notifications

Todo: change notification structure, once we have an notification structure/class that talks to the firestore side, we could easily intercept the number of notifications we have here instead of hard-coding it.

For example we would have a Notifications class and inside it we'd have a getter function that returns the length of the notifications' list. Because at the moment the notificationsList is hardcoded in the notification.ts but this should be moved.

* Header: make profile avatar clickable

Leading to the user's profile across the board.